news-aggregator/requirements.txt
Flask
feedparser
python-dotenv